casting, TotalMilliseconds is a double.

+++ SimpleBrowser/Browser.cs
@@ -166,7 +166,7 @@ namespace SimpleBrowser
         /// <summary>
         /// Sets the default browser timeout to this timespan. Must be > 0 milliseconds..
         /// </summary>
-        public void SetTimeout(Timespan timeout)=>Timeout=timeout.TotalMilliseconds;
+        public void SetTimeout(Timespan timeout)=>Timeout=(int)timeout.TotalMilliseconds;
 
         /// <summary>
         /// Gets or sets the browser culture.
